requirements.txt
anyio==4.0.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.7.22
chardet==3.0.4
exceptiongroup==1.1.3
googletrans==3.1.0a0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.17.3
httpx==0.24.1
hyperframe==5.2.0
idna==2.10
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3.post1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tornado==6.3.3
tzdata==2023.3
tzlocal==5.0.1
